requirements.txt
absl-py==2.1.0
accelerate==1.2.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.6
aiohttp-cors==0.7.0
aiohttp-retry==2.8.3
aiosignal==1.3.1
airportsdata==20241001
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
astor==0.8.1
astunparse==1.6.3
attrs==24.2.0
audioread==3.0.1
beautifulsoup4==4.12.3
blake3==1.0.2
blinker==1.7.0
blis==1.1.0
branca==0.8.0
buildozer==1.5.0
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.20.0
cloudpickle==3.1.0
colorama==0.4.6
colorful==0.5.6
compressed-tensors==0.8.1
comtypes==1.4.9
confection==0.1.5
contourpy==1.2.1
cycler==0.12.1
cymem==2.0.10
decorator==5.1.1
deepface==0.0.93
depyf==0.18.0
dill==0.3.9
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
dlib==19.24.6
docutils==0.21.2
einops==0.8.0
einops-exts==0.0.4
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et-xmlfile==1.1.0
face-recognition==1.3.0
face_recognition_models==0.3.0
facenet-pytorch==2.6.0
fastapi==0.115.6
fer==22.5.1
ffmpeg==1.4
filelock==3.16.1
fire==0.7.0
flamingo-pytorch==0.1.2
Flask==2.3.2
Flask-Cors==5.0.0
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.0.3
flatbuffers==24.12.23
folium==0.17.0
fonttools==4.53.1
fpdf==1.7.2
frozenlist==1.4.1
fsspec==2024.10.0
gast==0.6.0
gdown==5.2.0
geographiclib==2.0
geopy==2.4.1
gguf==0.10.0
google-api-core==2.23.0
google-auth==2.36.0
google-cloud-speech==2.28.0
google-cloud-vision==3.9.0
google-pasta==0.2.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.67.1
grpcio-status==1.67.1
gunicorn==23.0.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.7
imageio==2.36.1
imageio-ffmpeg==0.6.0
importlib_metadata==8.5.0
interegular==0.3.3
itsdangerous==2.1.2
Jinja2==3.1.3
jiter==0.8.2
joblib==1.4.2
JsonForm==0.0.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
JsonSir==0.0.2
keras==3.8.0
Kivy==2.3.0
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.7.0
Kivy-Garden==0.1.5
kiwisolver==1.4.5
langcodes==3.5.0
language_data==1.3.0
lark==1.2.2
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
llvmlite==0.43.0
lm-format-enforcer==0.10.9
lxml==5.3.0
lz4==4.3.3
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
mistral_common==1.5.1
ml-dtypes==0.4.1
moviepy==2.1.2
mpmath==1.3.0
msgpack==1.0.8
msgspec==0.19.0
mtcnn==1.0.0
multidict==6.1.0
murmurhash==1.0.11
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.60.0
numpy==1.26.4
nvidia-ml-py==12.560.30
ollama==0.4.6
openai==1.59.6
opencensus==0.11.4
opencensus-context==0.1.3
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.2
opt_einsum==3.4.0
optree==0.13.1
outlines==0.1.11
outlines_core==0.1.26
packaging==24.1
pandas==2.2.1
partial-json-parser==0.2.1.1.post5
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
playsound==1.3.0
pooch==1.8.2
preshed==3.0.9
proglog==0.1.10
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.21.1
proto-plus==1.25.0
protobuf==5.28.3
psutil==6.1.1
ptyprocess==0.7.0
py-cpuinfo==9.0.0
py-spy==0.4.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycountry==24.6.1
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pydub==0.25.1
pygame==2.6.1
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.1.2
pypiwin32==223
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
PySocks==1.7.1
python-dateutil==2.8.2
python-docx==1.1.2
python-dotenv==1.0.1
Python-EasyConfig==0.1.7
pyttsx3==2.98
pytz==2024.1
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
ray==2.40.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
resampy==0.4.3
Resource==0.2.1
retina-face==0.0.17
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
safetensors==0.4.5
schedule==1.2.2
scikit-learn==1.5.1
scipy==1.13.0
seaborn==0.13.2
sentencepiece==0.2.0
setuptools==75.4.0
sh==2.1.0
shellingham==1.5.4
sip==6.8.6
six==1.16.0
smart-open==7.1.0
sniffio==1.3.1
sounddevice==0.5.0
soundfile==0.12.1
soupsieve==2.6
soxr==0.3.7
spacy==3.8.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.13.0
SQLAlchemy==2.0.35
srsly==2.5.0
starlette==0.41.3
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
tf_keras==2.18.0
thinc==8.3.3
threadpoolctl==3.4.0
tiktoken==0.7.0
tokenizers==0.20.3
torch==2.2.2
torchaudio==2.5.1+cu118
torchvision==0.17.2
tqdm==4.67.0
transformers==4.46.2
twilio==9.3.2
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ultralytics==8.3.59
ultralytics-thop==2.0.13
urllib3==2.2.2
uvicorn==0.34.0
virtualenv==20.27.1
vllm==0.6.6.post1
wasabi==1.1.3
watchfiles==1.0.4
Wave==0.0.2
weasel==0.4.1
websockets==14.1
Werkzeug==3.0.1
wheel==0.44.0
wrapt==1.17.0
xyzservices==2024.9.0
yarl==1.12.1
zipp==3.21.0
